Only the 2 first output could be found
New output are supposed to be added to right of the last one which create
a list output.
So to find the third output it will be at the position of the sum of the
screen length of the two previous screen and not just the length of the
previous one.