A connected monitor doesn't mean it's used the laptop's one is always connected
When looking for connected monitor to move the workspace to it don't just check
for 'connected' value as the monitor can be plugin but not used which might
usually be the case with a laptop's screen when its closed.

Just keep trying to find the next screen (as active screen should be a "list")
until no one can be found.